Continued externalization of generators.
- Please ignore all the errors, I am not commiting all the classes I
have changed.

+++ PlotSquared/src/com/intellectualcrafters/plot/PlotBlock.java
@@ -1,9 +1,9 @@
 package com.intellectualcrafters.plot;
 
 public class PlotBlock {
-    public int id;
+    public short id;
     public byte data;
-    public PlotBlock(int id, byte data) {
+    public PlotBlock(short id, byte data) {
         this.id = id;
         this.data = data;
     }
